TST: Shifted a couple of tests around and moved them to pytest
